+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release 0.4.1
-- Using NSOperationQueue to manage tests
-- Changed how test groups run
+- Using NSOperationQueue to manage tests runs
+- Updated how test groups run
+- Handling failure in setUpClass/tearDownClass
+- Updated how shouldRunOnMainThread works
 
 # Release 0.3.19 (2008-06-15)
 - Fixed bug in Edit->Save crash (iPhone)